pyproject.toml
[tool.poetry]
name = "openclaw-mail-agent"
version = "1.0.0"
description = "Intelligent multi-account email management agent with AI-powered categorization"
authors = ["Redlex Gilgamesh"]
license = "MIT"
readme = "README.md"
packages = [{include = "openclaw_mail"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
ruff = "^0.4"

[tool.poetry.group.calendar.dependencies]
google-api-python-client = "^2.0"
google-auth-httplib2 = "^0.2"
google-auth-oauthlib = "^1.0"

[tool.poetry.scripts]
mail-tidy = "openclaw_mail.cli:tidy"
mail-digest = "openclaw_mail.cli:digest"
validate = "openclaw_mail.cli:validate"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
